Refactor Survey component into ROICalculation Survey, SurveyIntro, and SurveyQuestions components
